refactor quality_report
